Lab10

+++ IRCServer.cpp
@@ -499,7 +499,7 @@ IRCServer::getMessages(int fd, const char * user, const char * password, const c
 		token = strtok(NULL, blank);
 	}
 	
-	int lastMsgNum = atoi(vec[0].c_str());
+	long lastMsgNum = atol(vec[0].c_str());
 //	unsigned long lastMsgNum = stoul (vec[0],nullptr,0);
 	string room = vec[1];
